+++ src/game/nlcore/action/actionTypes.ts
@@ -31,7 +31,6 @@ export const SceneActionTypes = {
     action: "scene:action",
     setBackground: "scene:setBackground",
     sleep: "scene:sleep",
-    setTransition: "scene:setTransition",
     applyTransition: "scene:applyTransition",
     init: "scene:init",
     exit: "scene:exit",
@@ -39,20 +38,21 @@ export const SceneActionTypes = {
     setBackgroundMusic: "scene:setBackgroundMusic",
     preUnmount: "scene:preUnmount",
     applyTransform: "scene:applyTransform",
+    transitionToScene: "scene:transitionToScene",
 } as const;
 export type SceneActionContentType = {
     [K in typeof SceneActionTypes[keyof typeof SceneActionTypes]]:
     K extends typeof SceneActionTypes["action"] ? Scene :
         K extends typeof SceneActionTypes["sleep"] ? number | Promise<any> | Awaitable<any, any> :
             K extends typeof SceneActionTypes["setBackground"] ? [ImageSrc | ImageColor] :
-                K extends typeof SceneActionTypes["setTransition"] ? [ITransition | null] :
-                    K extends typeof SceneActionTypes["applyTransition"] ? [ITransition] :
-                        K extends typeof SceneActionTypes["init"] ? [] :
-                            K extends typeof SceneActionTypes["exit"] ? [] :
-                                K extends typeof SceneActionTypes["jumpTo"] ? [Scene] :
-                                    K extends typeof SceneActionTypes["setBackgroundMusic"] ? [Sound | null, number?] :
-                                        K extends typeof SceneActionTypes["preUnmount"] ? [] :
-                                            K extends typeof SceneActionTypes["applyTransform"] ? [Transform] :
+                K extends typeof SceneActionTypes["applyTransition"] ? [ITransition] :
+                    K extends typeof SceneActionTypes["init"] ? [Scene | string] :
+                        K extends typeof SceneActionTypes["exit"] ? [] :
+                            K extends typeof SceneActionTypes["jumpTo"] ? [Scene | string] :
+                                K extends typeof SceneActionTypes["setBackgroundMusic"] ? [Sound | null, number?] :
+                                    K extends typeof SceneActionTypes["preUnmount"] ? [] :
+                                        K extends typeof SceneActionTypes["applyTransform"] ? [Transform] :
+                                            K extends typeof SceneActionTypes["transitionToScene"] ? [IImageTransition, Scene | string | undefined, ImageSrc | ImageColor | undefined] :
                                                 any;
 }
 /* Story */