+++ src/game/nlcore/action/actions/conditionAction.ts
@@ -3,6 +3,7 @@ import type {Condition} from "@core/elements/condition";
 import {GameState} from "@player/gameState";
 import {LogicAction} from "@core/action/logicAction";
 import {TypedAction} from "@core/action/actions";
+import {Story} from "@core/elements/story";
 
 export class ConditionAction<T extends typeof ConditionActionTypes[keyof typeof ConditionActionTypes] = typeof ConditionActionTypes[keyof typeof ConditionActionTypes]>
     extends TypedAction<ConditionActionContentType, T, Condition> {
@@ -20,7 +21,7 @@ export class ConditionAction<T extends typeof ConditionActionTypes[keyof typeof 
         };
     }
 
-    getFutureActions(): LogicAction.Actions[] {
-        return [...this.callee._getFutureActions(), ...super.getFutureActions()];
+    getFutureActions(story: Story): LogicAction.Actions[] {
+        return [...this.callee._getFutureActions(), ...super.getFutureActions(story)];
     }
 }